After Update - Unity 5.20f3 -> 5.3.0f4
Everything is intact but collision are out of sorts now. Enemies are hit
and miss - which to some benefit it encourages a cautious play style
with fighting enemies since you need a good distance to jump away to
another platform if the enemy gets too close while "dodging" arrows -
but ultimately, this is obscene and I don't know how to fix it...

The Chest - can't collide with the arrows at all...had to change to
collision with Player to execute the ChestController code, an okay work
around...

Application.[insert member] is obsolete now in 5.3, uses "SceneManager"
now, but thank god I don't have to use it right now since Application.
is not completely deprecated.

I just hope nothing else goes wrong in making the next level...